Show reading progress as a whole percent
17.02% is noise on a progress bar. Rounding is held off the two ends it
would misreport: a reader with a page still to go is never shown 100%,
and one who has read a little of a long book is never shown 0%.

+++ src/ReadingTracker.Library/Entries/LibraryEndpoints.cs
@@ -245,7 +245,7 @@ public static class LibraryEndpoints
     /// Where the reader has got to. Derived from their latest session every time it is asked
     /// for, never stored, so it cannot disagree with the history it comes from.
     /// </summary>
-    private sealed record ProgressResponse(decimal Position, string Unit, decimal? PercentComplete)
+    private sealed record ProgressResponse(decimal Position, string Unit, int? PercentComplete)
     {
         public static ProgressResponse From(ReadingProgress progress) =>
             new(progress.Position, progress.Unit.ToString(), progress.PercentComplete);
